This seems like it would address #97. here's what it does:
1. change the `wordpress_id` field on the map table to be a varchar instead of an integer (updating the plugin from the wordpress repo will automatically update this, but users who update manually will need to do it manually)
2. when creating an object map between two items, plugin creates a temporary id value for whichever system does not have the item (ie a temporary wordpress id if we're on a salesforce pull action). This replaces the value of 0 that the plugin previously used before updating the database with the correct id value.

This way if the api fails, server goes down, whatever, that tmp id is unique down to the milisecond. The database key can stay, making sure duplicate ids do not get mapped and such, but in a worst case scenario there are rows that keep the temp id because the operation failed. They could at least be tracked down to when they happened and whether it was a push or pull that failed, by looking up that ID (the plugin does not yet facilitate showing all of these ids, and maybe should still provide an interface for that).