allow users to edit the error object map rows, in case they know what data they should contain.
there's also a noteworthy change in `salesforce_mapping` that filters the `$posted` array so it only contains the columns in the database table for object maps, plus the `action` parameter that we filter out later. have verified that the various `create_object_map` methods do not contain other parameters that we need to preserve for similar use to the `action` one.