Update class-object-sync-sf-wordpress.php to support user object on WP Multisite
Added lines 209-230 to check for WP Multisite installation. 

If it's a Multisite, then the 'user' object will be accessed using the 'wpdb->base_prefix' because the 'user' table is the same across all sites and just 'wpdb->prefix' does not work.

If it's not a Multisite, then the previous code executes.